Restructure repository layout
I made the repository more in-line with Arduino librarys in terms
of file and folder structure. I also added the README, license, and
boilerplate. Some files were renamed to fit this new structure.
Test "printf" messages were removed from the library. The example
file for the DualShock 4 was also made to be more user-friendly.